pyproject.toml
[tool.poetry]
name = "nonebot-plugin-mysticism"
version = "0.0.1"
description = "占卜辅助工具"
authors = ["Yan <1964649083@qq.com>"]
license = "Apache License 2.0"
readme = "README.md"

keywords = ["tarot", "nonebot", "nonebot2", "bot", "qq"]
homepage = "https://github.com/Yan-Zero/nonebot_plugin_mysticism"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.9.5"
nonebot2 = "^2.3.0"
nonebot-adapter-onebot = "^2.4.3"
pillow = "^10.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
